change the structure of the 'chat room'
