+++ app/src/view/main/chat/ChatMain.js
@@ -1,23 +1,52 @@
 import React, {Component} from 'react';
 import {inject, observer} from 'mobx-react';
 import Colors from 'react-native/Libraries/NewAppScreen/components/Colors';
-import {StyleSheet, Text, SafeAreaView, View, ScrollView, FlatList, TouchableOpacity} from 'react-native';
+import {
+    StyleSheet,
+    Text,
+    SafeAreaView,
+    View,
+    ScrollView,
+    FlatList,
+    TouchableOpacity,
+    Image,
+    ActivityIndicator
+} from 'react-native';
 import ChatMainHeader from './ChatMainHeader';
-import {ChatListMain} from "./list/ChatListMain";
-import {ChatListHeader} from "./list/ChatListHeader";
 import ChatRoomHeader from "./room/ChatRoomHeader";
 import ChatRoomMain from "./room/ChatRoomMain";
 import ChatRoomFooter from "./room/ChatRoomFooter";
 import {GROUPING_VIEW_STATUS} from "../../../constant/GroupingViewStatus";
 import GroupingCreationMain from "../grouping/creation/GroupingCreationMain";
 import {CHAT_VIEW_STATUS} from "../../../constant/ChatViewStatus";
+import {Icon} from "react-native-elements";
+import 'react-native-gesture-handler';
 
 @inject('chatStore')
 @observer
 class ChatMain extends Component {
-    // 컴포넌트가 만들어지고 첫 렌더링을 다 마친 후 실행되는 메소드입니다.
-    // 이 안에서 다른 JavaScript 프레임워크를 연동하거나,
-    // setTimeout, setInterval 및 AJAX 처리 등을 넣습니다.
+    state = {
+        data: [],
+        page: 0,
+        loading: false,
+    };
+
+    async componentWillMount() {
+        await this.fetchData();
+    }
+
+    fetchData = async () => {
+        const response = await fetch('https://randomuser.me/api?results=500');
+        const json = await response.json();
+        this.setState({data: json.results});
+    };
+
+    fetchChatRoomInfo = () => {
+        this.setState(
+            state => ({page: this.state.page + 1}, () => this.fetchData()),
+        );
+    };
+
     componentDidMount() {
     }
 
@@ -46,12 +75,12 @@ class ChatMain extends Component {
 
     render() {
         let view = (
-            <View style={styles.sex2}>
+            <View style={styles.main}>
                 <ChatMainHeader
                     onSearchViewBackButtonClicked={this.onSearchViewBackButtonClicked.bind(this)}
                 />
-                <ChatListMain/>
-                <FlatList
+                <AdArea/>
+                {/*<FlatList
                     onScrollBeginDrag={event => console.log('hello drag')}
                     data={DATA}
                     renderItem={({item}) =>
@@ -63,7 +92,33 @@ class ChatMain extends Component {
                         </Item>
                     }
                     keyExtractor={item => item.id}
-                />
+                />*/}
+                <View style={styles.test}
+                      padding={5}
+                      flexDirection={'column'}
+                      flex={1}>
+                    <FlatList
+                        style={styles.flex_list}
+                        data={this.state.data}
+                        keyExtractor={(x, i) => i}
+                        onEndReached={() => this.fetchChatRoomInfo()}
+                        onEndReachedThreshold={0}
+                        ListFooterComponent={() =>
+                            this.state.loading ? null : (
+                                <ActivityIndicator size="large" animating/>
+                            )
+                        }
+                        renderItem={
+                            ({item}) => (
+                                <Item
+                                    rowInfo={`${item}`}
+                                    onChatContentEnterButtonClicked={this.onChatContentEnterButtonClicked.bind(this)}
+                                />
+                            )
+                            // <ChatListRow rowInfo={} />
+                        }
+                    />
+                </View>
             </View>
         )
         if (this.props.chatStore.chatViewStatus === CHAT_VIEW_STATUS.ENTERING_CHAT_ROOM) {
@@ -131,29 +186,67 @@ const DATA = [
     },
 ];
 
-function Item({title, onChatContentEnterButtonClicked}) {
+function Item({rowInfo, onChatContentEnterButtonClicked}) {
     return (
         <TouchableOpacity
             onPress={() => onChatContentEnterButtonClicked()}
-            style={styles.item}
         >
-            <Text style={styles.title}>{title}</Text>
+            {/*<Text style={styles.title}>{rowInfo}</Text>*/}
+            <View style={styles.container}>
+                <Image style={styles.chatRoomImg}/>
+                <View flexDirection="column" flex={1} padding={5}>
+                    <View flexDirection="row">
+                        <Text
+                            style={styles.chatRoomTitle}
+                            numberOfLines={1}
+                            ellipsizeMode={'tail'}>
+                            Title
+                        </Text>
+                        <Text style={styles.numberOfParticipants}>5</Text>
+                        <Text style={styles.lastMessageTime}>
+                            12:00
+                        </Text>
+                    </View>
+                    <View flexDirection="row" alignItems={'center'}>
+                        <Text
+                            style={styles.lastMessage}
+                            numberOfLines={1}
+                            ellipsizeMode={'tail'}>
+                            {' '}
+                            lastMessage
+                        </Text>
+                    </View>
+                </View>
+            </View>
         </TouchableOpacity>
     );
 }
 
+function AdArea() {
+    return (
+        <>
+            <View style={styles.adArea}>
+                <Image
+                    style={styles.ad}
+                    source={{
+                        uri:
+                            'https://ssl.pstatic.net/tveta/libs/1281/1281372/38a98b76cb5b87ad613d_20200611113550630.jpg',
+                    }}
+                />
+            </View>
+        </>
+    );
+}
+
 const styles = StyleSheet.create({
-    sex2: {
-        flex: 1,
-        backgroundColor: 'blue',
-    },
     body: {
-        backgroundColor: Colors.primary,
+        backgroundColor: 'gold',
         flex: 1,
         height: '100%',
     },
-    container: {
+    main: {
         flex: 1,
+        backgroundColor: 'gold',
     },
     item: {
         backgroundColor: '#f9c2ff',
@@ -164,6 +257,78 @@ const styles = StyleSheet.create({
     title: {
         fontSize: 32,
     },
+    adArea: {
+        height: 70,
+        backgroundColor: 'black',
+    },
+    ad: {
+        height: 70,
+        borderRadius: 10,
+        margin: 5,
+    },
+    flex_list: {
+        flex: 1,
+        marginTop: 5,
+    },
+    test: {
+        backgroundColor: 'tomato',
+        flex: 1,
+        height: '100%',
+    },
+    container: {
+        flex: 1,
+        justifyContent: 'flex-start',
+        alignItems: 'center',
+        backgroundColor: '#FFF',
+        flexDirection: 'row',
+        height: 80,
+    },
+    flexlist: {
+        flex: 1,
+        marginTop: 5,
+    },
+    chatRoomImg: {
+        width: 65,
+        height: 65,
+        backgroundColor: '#eee',
+        alignItems: 'flex-start',
+        borderRadius: 25,
+        marginRight: 5,
+        borderWidth: 1,
+        borderColor: 'black',
+    },
+    chatRoomTitle: {
+        fontSize: 15,
+        fontWeight: 'bold',
+        flex: 1,
+    },
+    numberOfParticipants: {
+        fontSize: 14,
+        backgroundColor: '#eee',
+        borderRadius: 20,
+        padding: 3,
+        marginRight: 5,
+        marginLeft: 5,
+        height: 28,
+    },
+    lastMessageTime: {
+        color: 'gray',
+        fontSize: 12,
+        marginTop: 3,
+    },
+    lastMessage: {
+        color: 'gray',
+        flex: 1,
+    },
+    numOfNewMessages: {
+        color: 'white',
+        backgroundColor: 'red',
+        borderRadius: 15,
+        justifyContent: 'center',
+        padding: 3,
+        fontSize: 12,
+    },
+    imagesForTitleBar: {},
 });
 
 export default ChatMain;
