+++ app/src/view/main/chat/ChatMainHeader.js
@@ -1,13 +1,9 @@
 import React, {Component} from 'react';
 import {inject, observer} from 'mobx-react';
 import Colors from 'react-native/Libraries/NewAppScreen/components/Colors';
-import {StyleSheet, Text, View} from 'react-native';
-import {ChatListHeader} from "./list/ChatListHeader";
+import {Image, StyleSheet, Text, View} from 'react-native';
 
 class ChatMainHeader extends Component {
-    // 컴포넌트가 만들어지고 첫 렌더링을 다 마친 후 실행되는 메소드입니다.
-    // 이 안에서 다른 JavaScript 프레임워크를 연동하거나,
-    // setTimeout, setInterval 및 AJAX 처리 등을 넣습니다.
     componentDidMount() {
     }
 
@@ -18,13 +14,43 @@ class ChatMainHeader extends Component {
     ) {
     }
 
-    // 친구목록, 채팅, 모임찾기, 마이페이지
-
     render() {
         return (
             <View style={styles.body}>
-                {/*<Text>Chat</Text>*/}
-                <ChatListHeader/>
+                <View paddingRigth={8} paddingLeft={8}>
+                    <View style={styles.titleBar}>
+                        <Text style={styles.title}>채팅</Text>
+                        <View flex={1}/>
+                        <Image
+                            style={styles.titleImg}
+                            source={{
+                                uri:
+                                    'https://upload.wikimedia.org/wikipedia/commons/thumb/4/4c/Search_Noun_project_15028.svg/47px-Search_Noun_project_15028.svg.png',
+                            }}
+                        />
+                        <Image
+                            style={styles.titleImg}
+                            source={{
+                                uri:
+                                    'https://img.pngio.com/chat-icon-png-image-free-download-searchpngcom-chat-icon-png-715_657.png',
+                            }}
+                        />
+                        <Image
+                            style={styles.titleImg}
+                            source={{
+                                uri:
+                                    'https://upload.wikimedia.org/wikipedia/commons/thumb/3/35/Simple_Music.svg/600px-Simple_Music.svg.png',
+                            }}
+                        />
+                        <Image
+                            style={styles.titleImg}
+                            source={{
+                                uri:
+                                    'https://cdn4.iconfinder.com/data/icons/forgen-phone-settings/48/setting-512.png',
+                            }}
+                        />
+                    </View>
+                </View>
             </View>
         );
     }
@@ -37,7 +63,31 @@ const styles = StyleSheet.create({
         borderBottomColor: Colors.white,
         height: 60,
     },
+    container: {
+        flex: 1,
+        justifyContent: 'flex-start',
+        alignItems: 'center',
+        backgroundColor: '#FFF',
+        flexDirection: 'row',
+        height: 80,
+    },
+    titleBar: {
+        height: 40,
+        flexDirection: 'row',
+        alignItems: 'center',
+        backgroundColor: 'white',
+    },
+    title: {
+        fontSize: 20,
+        color: 'black',
+        flex: 1,
+        fontWeight: 'bold',
+    },
+    titleImg: {
+        width: 26,
+        height: 26,
+        marginRight: 8,
+    },
 });
-
 
 export default ChatMainHeader;
